Update Host.inf to avoid Emulator/Host CRASH in x64 linux platform
Emulater/Host call several functions that are defined in MdePkg and used EFIAPI calling convention, which is similar as ms_abi. Host's inf does not define the calling convention of GCC toolchain, which would use sysv_abi as gcc's default. They conflict to a crash when build in x64 linux platform. This bug is verified in Ubuntu 16.04/18.04/20.04 and Rocky Linux 9.2。